facturas_app/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
pytesseract==0.3.13
setuptools==75.8.0
Werkzeug==3.1.3
wheel==0.45.1
